requirements.txt
SQLAlchemy==2.0.36
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
